Align Drop on segments to actual user command.
Drop on segments was performed always using if exists,
this may cause inconsistency in cluster if for some reason table is
found on master but misses to find on segments. Patch now dispatchs DROP
with 'if_exists' only when it's specified by user. Otherwise, segments
will abort the transaction if the table being dropped is not found,
instead of just dropping table only on master.