Remove -Werror, replace with more specific -Wno-* options.
It's important to pay attention to compiler warnings, but -Werror isn't
buying us much at the moment. Firstly, there are a lot of warnings on modern
versions of gcc. I have to pass: "-Wno-error=unused-but-set-variable
-Wno-error=enum-compare -Wno-error=address -Wno-error=maybe-uninitialized"
to make it compile. More important than erroring out on warnings is to make
sure there are no warnings to begin with. Once we get to that state, any
warnings will stick out like sore thumb anyway, and you won't need -Werror
to make them more annoying.

We have a long way until we are 100% warning free, but most of the warnings
come from upstream code. They have been fixed in later versions of
PostgreSQL, so they will go away as we merge with upstream.