Fix FORCE QUOTE for integers and numerics.
Integers and numerics don't need quoting in CSV-format COPY, but they should
still be quoted if the FORCE QUOTE option is used.

We had memorized the incorrect output as expected output of the GPDB-added
parts at the end of the 'copy' regression test. Fix that, but also add
an explicit test case for this.