Fix and re-enable 'copy2' regression test.
Revert some unnecessary changes compared to upstream: there's no need for
ORDER BYs in queries, gpdiff.pl will mask out row order differences. That
needed some fixes to atmsort.pl though, to make it smarter about reordering
COPY TO STDOUT results. It used to only deal with the COPY (SELECT ...) TO
STDOUT variant, and particularly named tables (COPY .*test1 TO STDOUT). Make
it handle all COPY TO STDOUT commands. Also make it smarter about detecting
the end of COPY TO STDOUT output. In addition to -- or ERROR, also treat any
SELECT, UPDATE, CREATE etc. command as the end of result. In the passing,
remove the "-- copy_stdout" command from atmsort.pl, as it was unused.

No need to label the functions as immutable or with NO SQL or MODIFIES SQL
DATA, and update the error message in the expected output about
cross-segment access within functions.

One of the tests gives a different error than on upstream: an empty line
gives "invalid integer" error in upstream, but "missing data" on GPDB. I'm
not quite sure what's causing that, but both errors seem reasonable for that
case (the "missing data" is perhaps even better), so let's just memorize
that difference.