Fix and re-enable 'misc' regression test.
Remove unnecessary ORDER BY clauses that are not present in the upstream
version. Disable COPY BINARY test, as GPDB doesn't support BINARY copy.
Finally, adjust the error messages in the expected output to match what
you get nowadays.

DROP TABLE cleanup commands have been added to some other tests, so that
the extra user tables don't appear in the output of the 'misc' test. It's
not very nice of the 'misc' test to list all user tables, but this
nevertheless seems less likely to cause merge conflicts than ripping out
that part of the test altogether. There are still differences in the list
of user tables though.