Move tests for DECODE() from bugbuster to main test suite.
Looks like all the 'functional' test was testing was DECODE() expressions,
so rename the test, and move it to the main test suite. While we're at it,
clean up the test case to some extent. Fix the missing semicolons from some
queries, which caused the subsequent DROP to fail. Remove unused functions
and emp and dept tables. Add some comments explaining what each test does.
Remove use of multi-byte characters from test case, so that it works
regardless of encoding.