Use ssize_t for return value from send()
send() returns -1 in case of errors and thus we must use ssize_t rather
than size_t in order for the error check to be useful. A next pass over
this should be to check the actual errno of the triggered error to be
able to improve the error message.