Don't drop and re-create PL/pgSQL language as part of tests.
There seems to have been a philosophy for bugbuster tests to not only clean
up after test, but also try to drop any object that the test creates before
the test, so that the test case works if it is interrupted, and you re-run
it. We don't generally require that for test cases, as pg_regress always
creates a fresh 'regression' database to run the tests in.

In any case, dropping and recreating PL/pgSQL certainly doesn't seem
appropriate, because it's installed by default.