Replace pgsql-bugs@postgresql.org in backend translations.
We don't particularly care about these translations, as they haven't been
kept up-to-date in greenplum. The main reason for doing this is so that
they don't show up when doing "grep -R pgsql-bugs@postgresql.org .", to
find all the references that we really do care about.

Patch by Andreas Scherbaum