Fix pg_regress to only ignore NOTICES specified by the init_file.
pg_regress uses a tool called gpdiff.pl in order to compare diffs between
expected and actual result files in installcheck-good. pg_regress was hard
coded to always invoke gpdiff.pl to ignore NOTICE statements. Instead it would
be better to selectively ignore some NOTICE statements which are harmless and
which show up very often and compare the rest. This patch causes pg_regress to
invoke gpdiff.pl with an init_file which contains patterns that are to be
ignored.