Remove support for NLS.
There are no translations for Greenplum-specific strings, so it seems like
a disservice to allow anyone to build with --enable-nls.

The PostgreSQL translation files were a bit of a hodge-podge too; they
seemed to be from PostgreSQL 8.4.1, but we still carried the bunch of
incomplete translations that were removed in PostgreSQL 8.4.1. Let's not
pretend that there's any sense in them, and just remove all .po files. If
we want to make NLS to work properly in the future, we can bring them back,
using some sensible versions. Another option would've been to replace them
with the REL8_2_0 versions, which would possibly have made merging with
upstream slightly easier, as you wouldn't get merge conflicts on the
removed files, but those conflicts are easy to deal with. Also, by removing
the files completely, we don't need to decide what to do with the
references to pgsql-bugs@postgresql.org in them (we're trying to replace
all those references with bugs@greenplum.org).