Put back the out-function for PlannedStmt.
This partially reverts commit 4e01c159731b05f2a503c45938b0506a17414b5f.
The out-function for PlannedStmt is still needed, for debugging purposes,
even though we never try to read them back in from text format. There is
debug code in serializeNode that serializes a complete plan, and that
gives a WARNING if there is no out-function for PlannedStmt. (I'm not
convinced that code is doing much good, as it's only spewing stuff to the
log under DEBUG5, and at that level it will just get lost in the noise.
But that's a separate discussion). There is an out-function for
PlannedStmt in PostgreSQL 9.3 too, which is where we it was backported
from.

While we're at it, take the opportunity to move _outPlannedStmt to the
same location it's in in upstream, and remove spurious whitespace
differences.