Add feature to backup and restore catalog statistics.
A --dump-stats flag has been added to gpcrondump.py to dump the
pg_class table tuple counts and pg_statistic catalog statistics
to a SQL file, to later be restored manually or with gpdbrestore.

A --restore-stats flag has been added to gpdbrestore to restore
the dumped statistics.  Passing it the argument "include", or
passing it no argument, will restore the statistics in addition
to performing a normal restore.  Passing it the argument "only"
will restore statistics but will not restore anything else; if
any tables do not exist, using this option will skip restoring
statistics to those tables, print a warning for each, and then
continue restoring statistics to other tables.

If --dump-stats or --restore-stats is used in conjunction with
other flags that include or exclude certain tables or schemas,
the statistics dump or restore will be filtered in the same way.

Authors: James McAtamney and Jimmy Yih