Refactor gpmgmt1 test in bugbuster.
In the gpmgmt1 bugbuster test, the gpstate -f line makes a
connection to the template1 database which can cause the
other tests in gpmgmt1's parallel test group to fail.  An
example of this is when metadata_track tries to create
databases and fails as template1 is being accessed.

gpstate -f in gpmgmt1 bugbuster test has been removed.