Remove -msse4.2 flag from SUSE 11 compiler options.
This was put in some time ago as a work-around for detection of SSE 4.2
support. But we have pretty solid autoconf magic for that now, so that's
no longer needed. Moreover, that rock-solid autoconf magic was getting
confused by the way we pass compiler flags in CC rather than CFLAGS.