Update the 4.3.JSON file.
The 4.3.JSON file is used by gpcheckcat to figure out the foreign keys for
catalog tables. After making catalog changes, this file has to be generated by
running tidycat.pl and then copied over to gpMgmt/bin/gppylib/data/ directory.
If this file is not updated correctly to reflect the catalog changes it will
cause gpcheckcat to fail to identify foreign key inconsistencies.