Fix bug in IN-to-join conversion.
This bug was introduced in merge commit 1f4ad703, by a mishap in merging
the upstream changes to convert_IN_to_join. The constructed InInfoClause
entry was accidentally added to the list of in-clauses twice. That was
usually harmless, but if the in-clause was later pulled up from a subquery,
the code adjusts the varnos while merging the range tables of the subquery
and its parent, ran twice. In doing so, it adjusted the varno twice,
pointing it to wrong or non-existing range table entry.