Removing PrematureMDI release test and extra addref for caching object because new ccache version in gpos handles those cases.
PrematureMDI release test is removed because in gpos there is a new assert that checks if the cache object is valid for deletion and also ensure no memory leak.
CCache now take ownership of its object by adding ref and hence extra add ref in ORCA when we insert in CCache is removed.

resolved karthijrk/gp-orca#1
resolved #4
Signed-off-by: Karthikeyan Jambu Rajaraman <karthi.jrk@gmail.com>