Remove hardcoded information about pg_part_rule_def functions.
The hack was put in place, because prodataaccess values for these functions
are wrong in the catalogs. They should be marked as READS SQL DATA, but
they're marked as NO SQL. However, there are many more functions that are
similarly marked incorrectly, so we really should go through the pg_proc
catalog and fix all of them. But I'll leave that for another patch.

ORCA doesn't actually pay any attention to the prodataaccess field. It used
to to use it for optimization purposes at some point, but the code was taken
out because most functions are indeed mislabeled, and it was causing
problems.

This hack was subtly wrong anyway: the functions were marked as set
returning functions in the hard-coded structs, but they're not in fact
set-returning.

While we're at it, remove hardcoded OID of count() aggregate function, and
include it from pg_proc.h instead.