Fix quoting/escaping bugs in analyzedb.
This is a more comprehensive fix to the quoting/escaping issues that
lead to pull request #44. With this patch, analyzedb will no longer
choke on tables that contain funny characters in their names.

The behaviour is slightly different than that proposed in #44. If you
pass -t Foo.Bar, without any quoting, the server will resolve that using
the usual lowercase-smashing rule, and will search for "foo.bar", in
lowercase. To avoid that, table names should be quoted. This is the same
behaviour as e.g. the Postgres' reindexdb and vacuumdb tools have, so I
think that's reasonable. The -s option behaves differently though: case
is preserved.

Heikki Linnakangas with minor changes from Entong Shen.

Closes #187.