Pass parser state when resolving type name, like in upstream.
The user-visible effect is that you get error position in the original
query for some error messages, like the "type \"%s\" is only a shell"
error, if you try to use a shell type in CREATE TABLE, for example.

Not sure why this was different from upstream in GPDB, but seems prudent
to fix.