Reformat code, to match upstream more closely.
Adding a new 'return' like this in CheckAttributeType() is a bit risky, but
it seems better than re-indenting the whole function. Looking ahead at the
upcoming patches we're merging, we want to short-circuit all future checks
in this function too. In general, the reasoning is that any checks we might
want to perform on the attribute type should've been performed on the Query
Dispatcher already, and Query Executors should just follow orders.