Cosmetic changes to match upstream more closely.
It's unclear which upstream version our copy of this code is based on.
Looks like we have some of the changes from upstream commit b9527e98, but
not all. In any case, no upstream version has a variable called
'plannedstmt' here, all versions that have that variable call it 'pstmt'.