Fix buffer overflow in parsing boolean-type storage options.
For "checksum=on/off", we used the "on" or "off" value to allocate space
(i.e. 2 or 3 bytes), but then constructed a string that said "true" or
"false" (4 or 5 bytes). That lead to overflowing the allocation by a few
bytes, which produced warnings on assertion-enabled builds.

The integer parsing almost has the same bug. We use the original string's
length to allocate space, but we use %d to construct the value actually
stored there. That happens to not be a bug, because the original length is
always at least as large. This is a near-miss though: the original value
might be in hexadecimal format, i.e. "0x7fffffff", and its decimal
representation might be longer than the original. With 32-bit integers,
however, that happens to not be possible. as the extra "0x" in front of the
hex string provides enough padding. (i.e. strlen("0x7fffffff") >=
strlen("2147483647)"))

To fix, don't even try to count the space needed. Use a StringInfo instead.

In the passing, clean up accumAOStorageOpt function in some other ways:
remove the namelen argument to make it a bit more readable (this is hardly
performance-critical code), and don't split error message across lines
(makes it easier to grep for them if you need to find where an error
is coming from).