Improve the tests for gp_toolkit.
This brings in test queries from the "jetpack" test case in cdbfast (a
legacy Pivotal test suite that has not been open sourced yet), to cover
the gp_toolkit functionality that was not already covered by the bugbuster
tests. Move the test case from bugbuster into the main suite - this should
not be considered optional or auxiliary anymore.