Forbid DISTRIBUTED BY on an int2vector or aclitem column.
int2vector and aclitem don't have B-tree operators, making them not
mergejoinable. In the merge of the equivalence classes patch, I made
cdb_make_pathkeys_for_expr() more strict on that, and it now throws
an error if it's passed a datatype that doesn't have a mergejoinable
= operator. It used to return a path key with InvalidOid as the sort
operator, but that seems like a recipe for bugs elsehere.

To fix, mandate that all gpdb-hashable datatypes must be mergejoinable.
int2vector and aclitem are not, so forbid using them in DISTRIBUTED BY.

This is a backwards-incompatible change, but hopefully no-one is using
int2vector or aclitem in user-tables, at least not as a distribution key.