Handling unfinished sort due to QueryFinishPending
This fix prevents gpdb from a segmentation fault in case that: i) a SharedInputScan has a Sort child node, ii) sort spills, and iii) somebody requests to finish the execution during the last stage of the sort (i.e., QueryFinishPending = true). Currently, in the above scenario, SharedInputScan continues the execution and tries to access sort's result_tape, which is NULL, due to QueryFinishPending.

This commit makes a sort node to pretend that there are no results when QueryFinishPending flag is set to true. The commit also  creates a new Fault injector type and includes an ICG test.