Fix bug in checking whether a cached function is still valid.
When the combocids patch was backported from PostgreSQL 8.3 (c3983003), the
changes to change the cache key of compiled functions were left out for
some reason. Those changes are necessary: calling HeapTupleHeaderGetCMin()
on the pg_proc tuple will trip an assertion, if the tuple has been moved by
VACUUM FULL, and therefore HEAP_MOVED flags on the tuple are set. To fix,
backport the missing parts of the upstream combocid patch.