Fix confusion over opfamily used for a join clause in a window agg.
I'm not 100% the code is doing the right thing even with this: If the
pathkey's operator family doesn't match the join clause's operator family,
the output isn't truly ordered the same way. This is not an issue for the
case that I included in the regression test, where the join happens on a
constant, but I can't figure out whether this might be an issue with some
other query.