Remove comment referring to non-existent function.
I removed pathnode_copy_node() in an earlier commit, but missed this
comment. If someone needs to copy a Path in the future, they'll have to
come up with a way to do it at that point, but there hasn't been any need
for that so far. (copyObject() should work, but it makes a deep copy, while
the old pathnode_copy_node() function made a shallow copy.