Fix "variable not found in subplan target list" error.
Some test queries in the TINC/cdbfast test suite were failing with that
error, after merging the equivalence classes path. The problem was that
the Motion node's hash key an incorrect equivalence class member, one that
wasn't in the target list, instead of one that was.

To fix, make cdbpullup_findPathKeyExprInTargetList() return the actual Expr
that it finds, not just the EquivalenceClass containing it. The logic in
cdbpathlocus_get_partkey_exprs() that tried to find the right member was
not correct: it's not enough to check that the relids overlap.

This brings the code in question closer to the way it was before the
equivalence classes merge.