Change spelling of error message to match corresponding PostgreSQL message.
This eliminates the unnecessary difference in output depending on whether
a query is planned with ORCA or the Postgres planner. Which is good for
keeping regression tests simple.

Bump version number. Note that when GPDB is changed to use this version,
the error message in some GPDB regression test expected output files need
to be updated too.