Fix regression in pathkey reconstruction from EC merge
Finding the tle is a case for breaking the loop early but not finding
isn't a failure case. This elog() breaks the following query in the
gp_dqa testcase when found is properly initiated to false before the
loop.

select count(distinct c), count(distinct dt), i from dqa_t1 group by i;

When not initiated at all the variable was likely to not be false
from randomness and the test didnt fail.