Fix crash if a distribution hash expression contains a subplan.
We explicitly checked if the subplan was projection-capable, and added
the hash expression ot the subplan's target list if it was. But projecting
the hash expression is not optional, you get a crash if you don't do it. So
instead of doing it only when the subplan is projection-capable, insert a
Result node if needed.