Fix error-handling in dumpDatabaseDefinitions[ToDDBoost]() functions
Coverity pointed out that pszBackupFileName is referenced after free() in
the error path, in dumpDatabaseDefinition(). Testing shows that the
following fprintf() call will also segfault if the file could not be
created. In general, we should not try to continue running if one of these
files could not be created, so add exit_nicely() calls to the error paths.