Clean up the Vacuum and Analyze related structs.
Revert the VacAttrStats struct to look the same as in upstream. As a
consequence, we have to also revert the GPDB-specific changes to PostGIS
typanalyze functions. VacAttrStats is part of the public API for
type-specific typanalyze functions, so it's nice to revert that back the
way it is in the upstream, to make custom datatypes written for PostgreSQL
more easily portable to GPDB.

To my surprise, the VacAttrStats struct isn't currently used for anything,
so the changes to it were quite superfluous to begin with. That is to say,
we don't currently use type-specific typanalyze functions at all. That's
about to change, as part of a separate patch I'm working on to revert most
of analyze.c to use upstream code.

Also remove the unused VPgStatisticsStats struct, and the unnecessary
abstraction where VUpdateStats can be either VPgClassStats or
VPgStatisticsStats.