Simplify InitPlan elimination test case, and make its output more stable.
The test query used system tables pg_class and gp_configuration, which is
fine, except that the cost of scanning pg_class can vary widely, depending
on previous regression tests and the amount of relations created in them.
That made the output of the case a bit unstable; the planner might swap
the inner and outer sides of the nestloop join between pg_class and
gp_configuration, depending on the circumstances. That has been stable
so far, but started to cause grief on another patch I'm working on, which
changes the way ANALYZE works, because apparently it changed the statistics
collected for pg_class sligthly.

To fix, replace pg_class and gp_configuration in the test query with regular
tables, created specifically for the test case. While we're at it, simplify
the query a little bit, and also add a comment explaining what the purpose
of the test is. There doesn't actually seem to be any need to do EXPLAIN
here, the query produces wrong results if the bug is present, so get rid
of that too.

I tested this by reverting the changes that fixed the original bug this test
query was added for, and verifying that the new query also produces wrong
results with that.