Use upstream ANALYZE code.
This replaces most of the Greenplum code used to ANALYZE a relation with
the upstream implementation.

I did not adopt the upstream block-sampling method of collecting the
sample, however. That would difficult to make work with GPDB, because the
upstream code won't work as is with Append-only tables, and we would also
need to somehow dispatch the block-sampling to all the segments. So the
simplistic "SELECT ... WHERE random() < threshold" method is still used to
collect the sample. However, it's now just a straightforward read-only
query, we no longer create a table to hold the sample, which makes it a lot
cheaper for small tables.

This should make merging with upstream easier. This should also speed up
ANALYZE considerably, as we now collect the sample of rows in one query,
and calculate the stats for each column in-memory, instead of issuing
separate queries for each column.