Remove strlcpy.c from xlogdump build
strlcpy() isn't used in xlogdump and including it in the build cause
failures on platforms which supply strlcpy() in base. Remove source
and disconnect from Makefile.

This increases the diff in our version from upstream, however since
upstream is abandoned in favor of pg_xlogdump which was added to base
this is an Ok tradeoff to not have dead code in the tree.