Revert "Speed up gpinitsystem slightly."
This reverts commit a19e0e7e49733f5dddaaffe96a01185201455d5b. I had misread
the code, thinking that GP_HOSTADDRESS is a constant throughout the script.
It's not; it's set by the SET_VAR function, so the lookups were all looking
up different addresses. I did not notice that while testing on a single
host.

I still think the "sleep 2" is unnecessary, but I'm now afraid of touching
this because I don't have a multi-host setup at hand to test with. So revert
the whole commit .

Spotted by Jimmy Yih.