Make it more clear that test case contains an invalid COPY EOF-marker "\.  "
I wondered for a long time, why there are two copies of some DDL statements
in the .sql file, but only one copy in the .out file, and the test still
passes. Turns out that some of the DDL commands in the test file were inside
COPY data block. Instead of using valid-looking DDL as test data, put a
more explicit notice there. Perhaps putting a long block of valid queries
after a bogus "\.   " marker is a useful thing to test, but the point of the
test really should be explained explicitly.