Remove unnecessary test databases from schema_topology test.
The countrylist1 table was not used for anything either.

Makes it run a bit faster, I hope.

Now that some of the objects are created in 'regression' database, rather
than a newly created database, rename some objects to avoid clashing with
other bugbuster tests that use the same object names.