Remove checkstyle support.
I tried running this, and got a ton of errors. I don't think anyone's
used this for ages, so let's stop pretending that you might want to run
it. It might be a useful tool, as might some other style-checker or static
analysis tool, but this doesn't seem useful in its current state.