gp_truncate_error_log() function should first check for permissions on the master before dispatching the command to segments.
gp_truncate_error_log() function first dispatches the permissions check to the
segments and prints out the results before performing the check on the master.
This makes it hard to test this function if the cluster is configured with a
different number of segments than the test since different number of segments
produce different number of lines in the output. Performing the check first on
the master will cause the transaction to error out immediately and will be
easier to test.