Move version tests in mapreduce suite in under ignore blocks
Moves the Python and Perl version tests in under ignore blocks to
make the test less fragile. Testing for specific versions is not
ideal, if this is to be kept as a real test it should be rewritten
to look for interpreters satisfying minimum version requirements
rather then exact (although the value of that can be argued as
well). Keep the version output in the test though since it can be
good information when debugging where access is limited to the
regression.diffs file. Also fix a trivial typo in a comment while
in there.