Remove inaccurate assertion.
When DELETEing or UPDATEing an inherited table, some tables in the
inheritance tree might need an explicit Motion node to bring the targeted
tuples back to the segment where they reside, and some might not. The code
to build the plan handled that correctly, but this assertion incorrecly
assumed that it's all or nothing.

Remove the assertion, as it doesn't seem very useful in the first place.
The code that inserted the Motion nodes is just above the assertion, and
the assertion was basically just testing the same thing that the code
just did, and not some general invariant that should always hold.

Fixes issue #332