Validate pg_class pg_type tuple after index fetch.
While fetching the pg_class or pg_type tuple using index, perform sanity
check to make sure tuple intended to read, is the tuple index is
pointing too. This is just sanity check to make sure index is not broken
and not returning some incorrect tuple to contain the damage.