Revert the addition of 'rtoffset' to set_inner_join_references()
I'm not sure what the original problem was, because PostgreSQL is happy
without the rtoffset arg, and everything seems to work when I revert it.
I dug the original reproduction script referenced from when that change
was originally made and tested that with this revert, and it also worked
fine. Keeping the code as close to upstream as possible makes merging
easier.