Revert accidentally backported 8.4 changes from pg_trgm sources.
The trgm2int() function was backported from PostgreSQL 8.4, but never used.
And the TRGMINT() macro was removed as part of that backport, without
changing the callers.

This makes the code to compile, and the regression tests to pass, except
for the fact that GPDB doesn't support GIN, which produces an error.

Fixes issue #491, reported by @digoal.