Fix unused-initplans-removal machinery to handle SubPlans in PRECEDING clause.
remove_unused_initplans() didn't notice if there were references to a
subplan in the PRECEDING or FOLLOWING clause of a window function, because
plan/expression walkers didn't visit the PRECEDING/FOLLOWING clauses. Fix
the walkers.

We already had a test case for this in the regression suite, in
gp_optimizer, but instead of fixing the bug we had memorized the internal
ERROR message in the expected output. Oopsie.