Fix the rejected row count reported by single row error handling in COPY.
When we use error log, if the master finds a badly formatted data row, it
increments the rejected row count and then sends the row to the segment so that
it can be stored in the error log file. On the segments, the row gets parsed
again and the segment increments the reject count again.  For reporting the
total number of rejected rows to the user we sum up the rejected row count from
the master and the segments. We need to ignore the count from the master
because it will be included in the reject count from the segments.

Also includes additional cleanup and typo fix of variable name by Heikki
Linnakangas.