Automatic casting for partition keys
Under multi-level partitioning, GPORCA generates partition boundaries
based on the partition definition. If the data types of the partition
columns (aka partition keys) are not comparable directly with the
boundary values, a casting is required to cast the partition keys to
the data types of boundary values.

We also cleanup some deadcode in CTranslatorDXLToPlStmt.h.

[#110377632]