Remove incorrect assumption that reindex cannot happen within a transaction block.
ReindexRelationList is a GPDB specific function that wraps each
reindex table operation within a separate transaction.  If the table
being reindexed is a partitioned table, each partition is reindexed in
a separate transaction.  Now that we have removed the assertion,
refactor this function to commit and start top level transcation,
rather than leaving this burden on callers.

Fixes #567.