Move gpcheckcat symlink installation to install dir from source tree
Commit 41dfd82 moved gpcheckcat from bin/lib to bin/ and installed a
symlink to maintain backwards compatibility. However, installing and
copying the symlink in the source tree makes Git complain that it's
an untracked file. Rather than adding it to .gitignore we might as
well do this in the installation directory since that's where we want
to maintain backwards compatability anyways.

Cleaned up a few comments around the installation as well while in
there.