Fix error log test for external table.
The cursor test for external table is non deterministic. The query errors out
either during DECLARE cursor or during FETCH. If the query errors out during
DECLARE and we perform a FETCH after the DECLARE an extra line of error is
produced in the output stating that the transaction has been aborted. In the
case where FETCH errors out this error is not produced in the output. This
commit adds a separate expected answer file to handle the two cases.