Synchronize the OIDs of shell operators created for commutator/negator.
If you do CREATE OPERATOR, with a commutator or negator operator that
doesn't exist yet, the system creates a "shell" entry for the non-existent
operator. But those shell operators didn't get the same OID in all segments,
which could lead to strange errors later. I couldn't find a test case
demonstrating actual bugs from that, but it sure seems sketchy. Given that
we take care to synchronize the OID of the primary created operator, surely
we should do the same for all operators.