Give a WARNING if an object is allocated an OID in segment.
A common source of bugs has been that an object gets assigned a different
OID in the master and in segments. A segment should normally never have to
allocate an OID (for catalog objects) on its own, all OIDs should be
allocated in the master, and sent over to the segments. To make such bugs
easier to catch, add a WARNING if an OID is allocated in a segment.

There were some DEBUG1 elogs for the same thing in place already, but the
list of catalogs that need synchronized OIDS wasn't up-do-date, and this new
place for the elog() is less invasive anyway.