Enhance partition_locking test case to also check locks in segments.
While hacking in this area, at one point I broke the code so that a
partition was not correctly in a segment as it should've been, even though
it was locked in the master. Fortunately, I noticed that in manual testing
before committing, but to prevent such bugs from slipping in in the future,
enhance the 'partition_locking' test to catch that kind of bugs.