add -n for sort command when initsystem compare the primary and mirror order
since I get error when initsystem if the segment number >8
[FATAL]:-mismatch between content id and primary content id

I check that this is because, when matching mirror with primary has different order for some index.
e.g. I print the sort result:
primary is 0,10,11,12,13,1,14,15...
mirror is 0,10,11,12,13,14,1,15....

I believe it make more sense that we use sort -n , so I changed the code, and init system successfully.

Patch submitted by @byxiangfei