Revert unnecessary changes vs. upstream.
Revert the change to use binary search instead of linear search, for
searching the type array. Bootstrapping is hardly performance critical, and
the array isn't that large anyway.

Remove "bigint" from the array. I'm not sure what it was added for originally,
but it's not present in the upstream, and initdb seems to work without it.