Remove WATCH_VISIBILITY_IN_ACTION debugging aid.
It might be useful in debugging, but it got pretty badly in the way while
merging with PostgreSQL 8.3. We could fix it, of course, but on balance,
I don't think it's worth the effort. It's going to be a maintenance burden
going forward too, as the WATCH_* calls are scattered all over the
visibility checking code. If we need debugging code like that, we should
find a less invasive way to implement it, or submit the mechanism to
upstream so that we wouldn't need to maintain it as a diff.