Remove a lot of unnecessary setup steps from qp_misc regression test.
There were a lot of unused tables and functions, and chaff like comments
that are not needed for the actual tests in the file. At first glance,
some of the things seemed marginally useful to test on their own right,
like loading data with non-ASCII characters in it, but all the setup
stuff was in a large ignore-block, so any failures there would go unnoticed
anyway.

Removing unnecessary stuff is a virtue of its own, but this also speeds up
the test nicely.