Avoid infinite loop on error.
While hacking, I ran into the "Expected a CREATE for file-system object
name" error. But instead of printing the error, it got into an infinite
loop. smgrSortDeletesList() elog(ERROR)'d out of the function, while it
was in the middle of putting the linked list back together, leaving
pendingDeletes list corrupt, with a loop. AbortTransaction() processing
called smgrIsPendingFileSysWork(), which traversed the list, and it got
stuck in the loop.

To avoid that, don't leave the list in an invalid state on error. I don't
know why I ran into the error in the first place, but that's a different
story.