Remove secondary clutter from concourse pipeline.
Removed the secondary from the group all to cut down on the clutter
in the all view of Concourse. Secondary has been renamed to platforms
because its purpose is to build and test on different platforms, but
different from the pipeline all, which is used for the ICG.

Non-centos platform builds and tests were moved to the platforms group.

Fixed a source dependency to build_gpdb_centos6 and also orca_centos5_release.