[#118865569] Fix bug of wrong Broadcast with Window Function
Bump GPORCA version to 1.637

When there is a Gather on top of the window function
(CPhysicalSequenceProject), there cannot be a broadcast below that
operator.

The fix got two parts. First, fix-up the CDistributionSpecAny to
remember the EOperatorId of who requests it. This will help the
reasoning later on in the CPhysicalSequenceProject to decide what kind
of required distribution spec to request from its child.

Second, we fix up the CPhysicalSequenceProject to avoid generating a
CDistributionReplicated if a Gather motion generated an
CDistributionSpecAny.

Besides those, we also cleanup the testing framework:
- rename Fst --> Actual and Snd --> Expected in UnitTest.
  This helped a lot reading what's expected and what actual plan
  produced in the gporca unit tests.
- refactored the CICGTest a little to separate window function tests out
- fix minidump option `-d` to have same configuration as `-U`