Lower minimum work_mem to 64kB, also on non-assert-enabled builds.
A new regression test was added in the upstream that does work_mem='64kB',
which used to fail without --enable-cassert. There's little reason to not
allow a small value here, so let's just allow it. It doesn't mean that
anyone should use such a small value in production, but that's just a
matter of "don't do that then".

The minimum was changed later in the upstream, too, to be 64 kb regardless
of BLCKSZ.