Create naive code-generated version of ExecQual
This commit generates code for a naive code-generated version of ExecQual that simply fallbacks to regular ExecQual.
The code-generated ExecQual is enrolled in the existing codegen framework and is called only in ExecScan.
After the implementation of the function (i.e., generate code for expression evaluation), we can replace all calls to regular ExecQual with call_ExecQual macros.