Remove leftover from backporting that breaks compilation of xml2
The tuplestore in xpath_table() is created with random access even
though it might not require it, the scaffolding for dealing with
relaxing this (commit 05bba3d176e0adc1a032d5c8c6ea2a7622c7dd0d) came
the in 8.4 cycle and is yet to be merged. For now drop down to always
requesting random access awaiting the merge of the 8.4 patches.