Mask trivial differences in EXPLAIN output in bfv_index test.
pg_regress uses gpdiff.pl to mask differences in costs etc. of EXPLAIN
output, so that trivial changes in the plan output don't cause a
regression failure. That mechanism didn't work for this EXPLAIN query,
because the "explain" and "select" were on different lines. Remove the
newline, to make gpdiff.pl recognize the query as an EXPLAIN query, to
make the test case less fragile.