Add missing field to readRangeTblEntry, improve comment in fast version.
The readRangeTblEntry() function was missing the line for 'pseudocols'
field. I'm surprised the function worked at all, I thought the read
functions don't work if there are any extra fields. Maybe they're more
forgiving if it's last field that's missing. In any case, seems like an
oversight. It doesn't matter in practice, as the pseudocols field is only
used during planning, and we don't serialize nodes at that stage. Rule
definitions are serialized before planning, and for the transfer between
QD and QEs, we use the 'fast' versions of these functions.

In the 'fast' version, the 'pseudocols' is missing from both the out
function and the read function. That seems intentional, so add a comment
about it.