Silence WARNINGs about missing out function with DEBUG5
If log_min_messages = 'debug5', we call the text output function for
everything that's serialized to be sent from QD to QEs. I added a new
node type, QueryDispatchDesc, in commit 41478b89, but I only gave it
the "fast" output function, not the text version. That produced a
warning about missing output function, with log_min_messages='debug5'.

This fixes the failure of qp_executor test, reported by Omar Arap.