Remove duplicated bitgetpage() function.
The one in nodeBitmapHeapScan.c is what the upstream has. The one in
execBitmapHeapScan.c was copied from it in GPDB. No need for the
duplication.

It's not clear to me why we have the execBitmapHeapScan.c file at all,
why not just use all the functions in nodeBitmapHeapScan.c. But I'll
leave investigating that for another day.